==> 2-descriptive-outcomes/wasting/15_pool_birth_RRs.R <==


rm(list=ls())
source(paste0(here::here(), "/0-config.R"))
source(paste0(here::here(), "/0-project-functions/0_clean_study_data_functions.R"))
source(paste0(here::here(), "/0-project-functions/0_risk_factor_functions.R"))


#Load data
dfull <- readRDS(paste0(here::here(),"/results/rf results/raw longbow results/results_early_wast_2020-06-03.RDS"))
head(dfull)

d <- dfull %>% filter(type=="RR", ci_lower != ci_upper, agecat=="0-6 months")

RMAest <- d %>% group_by(intervention_variable, agecat, intervention_level, baseline_level, outcome_variable) %>%
  do(poolRR(.)) %>% as.data.frame()
RMAest


RMAest_FE <- d %>% group_by(intervention_variable, agecat, intervention_level, baseline_level, outcome_variable) %>%
  do(poolRR(., method="FE")) %>% as.data.frame()
RMAest_FE
